🔥 Remove prefixed devices
