+++ examples/device/CMakeLists.txt
@@ -6,5 +6,3 @@
 
 add_subdirectory(c)
 add_subdirectory(cxx)
-add_subdirectory(unprefixed_c)
-add_subdirectory(unprefixed_cxx)
