--- examples/device/device_shifted.h.in
@@ -1,52 +1,0 @@
-/*------------------------------------------------------------------------------
-Part of the MQSS Project, under the Apache License v2.0 with LLVM Exceptions.
-See https://llvm.org/LICENSE.txt for license information.
-SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-------------------------------------------------------------------------------*/
-
-#pragma once
-
-#ifdef QDMI_PREFIX
-#undef QDMI_PREFIX
-#endif // QDMI_PREFIX
-#define QDMI_PREFIX @prefix@
-
-#define QDMI_Job_impl_d @prefix@_QDMI_Job_impl_d
-#define QDMI_Site_impl_d @prefix@_QDMI_Site_impl_d
-#define QDMI_Operation_impl_d @prefix@_QDMI_Operation_impl_d
-#define QDMI_query_get_sites_dev @prefix@_QDMI_query_get_sites_dev
-#define QDMI_query_get_operations_dev @prefix@_QDMI_query_get_operations_dev
-#define QDMI_query_device_property_dev @prefix@_QDMI_query_device_property_dev
-#define QDMI_query_site_property_dev @prefix@_QDMI_query_site_property_dev
-#define QDMI_query_operation_property_dev @prefix@_QDMI_query_operation_property_dev
-#define QDMI_control_create_job_dev @prefix@_QDMI_control_create_job_dev
-#define QDMI_control_set_parameter_dev @prefix@_QDMI_control_set_parameter_dev
-#define QDMI_control_submit_job_dev @prefix@_QDMI_control_submit_job_dev
-#define QDMI_control_cancel_dev @prefix@_QDMI_control_cancel_dev
-#define QDMI_control_check_dev @prefix@_QDMI_control_check_dev
-#define QDMI_control_wait_dev @prefix@_QDMI_control_wait_dev
-#define QDMI_control_get_data_dev @prefix@_QDMI_control_get_data_dev
-#define QDMI_control_free_job_dev @prefix@_QDMI_control_free_job_dev
-#define QDMI_control_initialize_dev @prefix@_QDMI_control_initialize_dev
-#define QDMI_control_finalize_dev @prefix@_QDMI_control_finalize_dev
-
-#include "qdmi/device.h" // IWYU pragma: export
-
-#undef QDMI_Job_impl_d
-#undef QDMI_Site_impl_d
-#undef QDMI_Operation_impl_d
-#undef QDMI_query_get_sites_dev
-#undef QDMI_query_get_operations_dev
-#undef QDMI_query_device_property_dev
-#undef QDMI_query_site_property_dev
-#undef QDMI_query_operation_property_dev
-#undef QDMI_control_create_job_dev
-#undef QDMI_control_set_parameter_dev
-#undef QDMI_control_submit_job_dev
-#undef QDMI_control_cancel_dev
-#undef QDMI_control_check_dev
-#undef QDMI_control_wait_dev
-#undef QDMI_control_get_data_dev
-#undef QDMI_control_free_job_dev
-#undef QDMI_control_initialize_dev
-#undef QDMI_control_finalize_dev
